Add files via upload
1. Outline of the project
The goal of this project is to use the tree graph as an inference (prediction) tool. You
are provided with a dataset of observations x1, . . . , xn, n = 100. Each observation xi is
a 2-vector containing a value belonging to R and a label belonging to {0, 1}. We assume
here that the two labels represent two different groups of observations and the final goal is
to build a decision tree that will assign a label (i.e. a group) to a new observation ˜x with
unknown label. You will use the CART (Classification and Regression Tree) Algorithm to
solve this problem. This algorithm aims to build a binary decision tree that assign a label
to ˜x. Every vertex of a binary decision tree is such that it has exactly 0 or 2 children and
every vertex contains a threshold value. To classify a new observation ˜x, we give ˜x as an
input to the classification tree and check if said observation is smaller than the threshold
value contained in the root node. If it is the case, we travel to the first child node, else to
the second one. We repeat this process until we get to a node (let say Nt) with no children
(called a leaf). Then, we predict the label of ˜x by taking all the datapoints in {x1, . . . , xn}
such that they belong to the node Nt and use as an estimator of the label of ˜x the most
common label value for all the datapoints in {x1, . . . , xn} such that they belong to such
node. The goal is, given x1, . . . , xn to find a set of threshold values that will give some
predictive power to the binary decision tree.
You can use Python, R or C++ to complete the project and should submit through
moodle only a documented source code file.
2. Step by step
1. You are provided a .csv file, open it with Python/R/C++. The first column (V1) rep-
resents the value of each datapoint while the second column (V2) represents its label.
2. You will build the decision tree T now. You will use the so-called Gini impurity to
assess the quality of a split of the dataset, split the data in two subsets then find
another good split for each one of these subsets and so on. We denote the label of
xi as l(xi) and the value of xi as v(xi). At the first step, you need to determine the
value for the first split c1 ∈ R. First check if the number of observations you have in
1


the first node is larger that some threshold α with α = 8 to avoid overfitting. α acts
here as a parameter imposing that the number of leaves won’t be too large and its
value (8) is given arbitrarily. Let define
P1 = |{xi | v(xi) < c1}| and Q1 = |{xi | v(xi) > c1}| and p1 =
P
{xi|v(xi)<c1} l(xi)
P1
and q1 =
P
{xi|v(xi)>c1} l(xi)
Q1 . Here p1 and q1 represent the proposition of label 1
observations in the subsets made of the observations respectively smaller and larger
than c1. You will chose c1 such that it minimises
P1p1(1 − p1) + Q1q1(1 − q1),
the so-called Gini impurity. Then, iterate : check if |{xi | v(xi) < c1}| > α and de-
termine in the same way as before c2 for the subset of observations {xi | v(xi) < c1},
check if |{xi | v(xi) > c1}| > α, then determine c3 for for the subset {xi | v(xi) > c1}
and so on up until every leaf (node without children) is such that it contains at max-
imum α observations. You need to store all the threshold values c1, . . . , ck in a tree
data structure that should allow you, given a node Nk to get both the parents and
children nodes of Nk. You can implement the tree data structure how you see fit.
If you want to do it properly, the best way is to use object oriented programming
but, since the problem at hand here is very simple it is also acceptable to store the
tree in a matrix M of size k × 4. With such a structure, nodes are identified with
integers i ∈ {1, . . . , k} and the node i, 1 ≤ i ≤ k is such that the i-th line of M,
denoted Mi satisfies Mi = (ci, j, l, m) with ci the value of the threshold contained in
the node, j ∈ {1, 2, . . . , k} the parent node, l ∈ {1, 2, . . . , k} the first child node and
m ∈ {1, 2, . . . , k} the second child node. Of course, no split will occur in a leaf by
constricution so you can put for instance ci = 0 arbitrarily for any leaf (or entirely
discard such nodes and don’t even put them in N ). What is expected of you at this
point is to provide this classification tree T , implemented through the matrix M or
any other way you like that allows to make prediction about new unlabelled data.
Remark: In practice, the threshold α should be determined by cross-validation by
looking at the misclassification error to get an optimal classification tree but here,
for the sake of simplicity this α is given to you as a lower bound on the number of
observations in a node.
3. Now, given new observations ˜x1, ˜x2 and ˜x3 such that they have values v(˜x1) = 1.2,
v(˜x2) = −4 and v(˜x3) = 3, write a short script using your decision tree T to predict
the labels of ˜x1, ˜x2 and ˜x3. To do that, first determine in which leaf (node without
children) v(xi), i = 1, 2, 3 belongs, then predict its label by the most common label


value observed in the subset of observations present in said leaf.
3. Grading
There is no need to be extra efficient in your implementation of both the data tree structure
and the algorithm that finds the optimal split ci, i = 1, . . . , k even if a clean and efficient
implementation will be rewarded. You will be mainly evaluated on two main points:
(1) Is the tree T that your algorithm built a proper tree? I.e., can we recover for every
node its parent node, children nodes and the value ci of the split it contains?
(2) Is the tree T a proper prediction tree allowing to make sensible prediction regard-
ing the label of the new points ˜x1, ˜x2 and ˜x3? I would encourage you to check
that the predicted label you get in the end make sense with respect to the dataset
{x1, . . . , xn}